pyproject.toml
[tool.poetry]
name = "accounting manager"
version = "0.1.0"
description = ""
authors = ["Ilya Verbilo <verbiloilyaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
django-filter = "^22.1"
celery = {extras = ["redis"], version = "^5.2.7"}
redis = "^4.3.4"
django-celery-beat = "^2.4.0"
django-environ = "^0.9.0"
flake8 = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
